requirements.txt
django-textify
django-massmedia
django-categories
